backend/app/main.py
"""
AI Need Human - FastAPI Backend
Main application entry point with CORS and route configuration
"""

from fastapi import FastAPI, Depends
from fastapi.middleware.cors import CORSMiddleware
from fastapi.staticfiles import StaticFiles
from fastapi.responses import FileResponse
from contextlib import asynccontextmanager
from datetime import datetime
import os
from dotenv import load_dotenv
from supabase import Client

# Load environment variables
load_dotenv()

# Import routers (will be created)
from routers import auth, users, services, bookings, messages, payments
from services.supabase import get_supabase


@asynccontextmanager
async def lifespan(app: FastAPI):
    """Application lifespan events"""
    # Startup
    print("🚀 Starting AI Need Human API...")
    print(f"📍 Environment: {os.getenv('ENVIRONMENT', 'development')}")
    yield
    # Shutdown
    print("👋 Shutting down AI Need Human API...")


# Initialize FastAPI app
app = FastAPI(
    title="AI Need Human API",
    description="First marketplace where AI agents can hire humans through MCP",
    version="1.0.0",
    docs_url="/docs",
    redoc_url="/redoc",
    lifespan=lifespan,
)

# CORS Configuration
CORS_ORIGINS = os.getenv("CORS_ORIGINS", "http://localhost:3000").split(",")

app.add_middleware(
    CORSMiddleware,
    allow_origins=CORS_ORIGINS,
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Mount static files
if os.path.exists("static"):
    app.mount("/static", StaticFiles(directory="static"), name="static")

# Demo page
@app.get("/demo", tags=["Demo"])
async def demo_page():
    """Live demo page"""
    return FileResponse("static/demo.html")

# Health check endpoint
@app.get("/", tags=["Health"])
async def root():
    """API health check"""
    return {
        "status": "healthy",
        "message": "AI Need Human API is running",
        "version": "1.0.0",
        "docs": "/docs",
        "demo": "/demo"
    }


@app.get("/health", tags=["Health"])
async def health_check():
    """Detailed health check"""
    return {
        "status": "healthy",
        "database": "connected",
        "mcp": "available",
        "blockchain": "connected"
    }


@app.get("/api/stats", tags=["Stats"])
async def get_stats(supabase: Client = Depends(get_supabase)):
    """Get live platform statistics for landing page"""
    try:
        # Count users
        users = supabase.table("profiles").select("id", count="exact").execute()

        # Count services
        services = supabase.table("services").select("id", count="exact").eq("is_active", True).execute()

        # Count bookings
        bookings = supabase.table("bookings").select("id", count="exact").execute()

        # Count messages
        messages = supabase.table("messages").select("id", count="exact").execute()

        return {
            "users": users.count if users.count else 10,
            "services": services.count if services.count else 20,
            "bookings": bookings.count if bookings.count else 8,
            "messages": messages.count if messages.count else 12,
            "last_updated": datetime.now().isoformat()
        }
    except Exception as e:
        # Fallback to demo data if database query fails
        return {
            "users": 10,
            "services": 20,
            "bookings": 8,
            "messages": 12,
            "last_updated": datetime.now().isoformat(),
            "note": "Demo data (database connection error)"
        }


# Include routers
app.include_router(auth.router, prefix="/auth", tags=["Authentication"])
app.include_router(users.router, prefix="/users", tags=["Users"])
app.include_router(services.router, prefix="/services", tags=["Services"])
app.include_router(bookings.router, prefix="/bookings", tags=["Bookings"])
app.include_router(messages.router, prefix="/messages", tags=["Messages"])
app.include_router(payments.router, prefix="/payments", tags=["Payments"])


if __name__ == "__main__":
    import uvicorn

    host = os.getenv("API_HOST", "0.0.0.0")
    port = int(os.getenv("API_PORT", "8000"))

    uvicorn.run(
        "main:app",
        host=host,
        port=port,
        reload=True,
        log_level="info"
    )
